chore: make the type-ignore and migration-parent ratchets count what they claim [order J-ratchets]
