+++ .type-ignore-baseline
@@ -1,1 +1,1 @@
-62
+61
