+++ tests/integration/migration_helpers.py
@@ -8,14 +8,17 @@ isolated PostgreSQL databases:
 
 The shared ``migration_db`` fixture lives in ``conftest.py`` (pytest auto-discovers
 fixtures from conftest without explicit imports, avoiding F811 lint errors).
+
+The row-touching helpers (``seed_tenant``, ``column_exists``, ``get_column_info``)
+live in ``tests/helpers/migration_helpers.py`` instead — that directory is inside
+the repository-pattern guard's scan set, and shared test substrate belongs where a
+guard can see it.
 """
 
 from __future__ import annotations
 
 import os
 import re
-
-from sqlalchemy import text
 
 
 def parse_postgres_url() -> tuple[str, str, str, int] | None:
@@ -64,43 +67,3 @@ def run_alembic_downgrade(db_url: str, target_revision: str) -> None:
     from alembic import command
 
     _run_alembic_command(db_url, command.downgrade, target_revision)
-
-
-def column_exists(engine, table: str, column: str) -> bool:
-    """Whether ``table.column`` exists (information_schema probe).
-
-    Shared by migration tests that assert a column is added/dropped, replacing
-    per-test hand-rolled copies of the same query.
-    """
-    return get_column_info(engine, table, column) is not None
-
-
-def get_column_info(engine, table: str, column: str) -> tuple[str, str | None] | None:
-    """``(is_nullable, column_default)`` for ``table.column``, or None if absent."""
-    with engine.connect() as conn:
-        row = conn.execute(
-            text(
-                "SELECT is_nullable, column_default FROM information_schema.columns "
-                "WHERE table_name = :t AND column_name = :c"
-            ),
-            {"t": table, "c": column},
-        ).first()
-    return (row[0], row[1]) if row else None
-
-
-def seed_tenant(engine, tenant_id: str, *, subdomain: str) -> None:
-    """Insert a minimal tenant row on a migration-managed schema.
-
-    The tenant INSERT is identical across migration tests; centralizing it here
-    keeps new migration tests from re-copying it (each test still seeds its own
-    domain-specific child rows).
-    """
-    with engine.connect() as conn:
-        conn.execute(
-            text(
-                "INSERT INTO tenants (tenant_id, name, subdomain, created_at, updated_at) "
-                "VALUES (:tid, :name, :sub, NOW(), NOW())"
-            ),
-            {"tid": tenant_id, "name": f"{tenant_id} tenant", "sub": subdomain},
-        )
-        conn.commit()
